+++ src/square.js
@@ -1,0 +1,12 @@
+function Square(props) {
+  return (
+    <button
+      className="square"
+      onClick={props.onClick}
+    >
+      {props.value}
+    </button>
+  )
+}
+
+export default Square;
